requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
alembic>=1.13.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
pydantic>=2.0.0
jinja2>=3.1.0
aiofiles>=23.0.0
openpyxl>=3.1.0

# Alternatif: Flask tercih ederseniz yukarıdakileri silerek bunları kullanın:
# flask>=3.0.0
# flask-login>=0.6.0
# flask-sqlalchemy>=3.1.0
# flask-migrate>=4.0.0
# flask-wtf>=1.2.0
# bcrypt>=4.0.0
